+++ eelbrain/table.py
@@ -1,4 +1,12 @@
 # Author: Christian Brodbeck <christianbrodbeck@nyu.edu>
 """Create tables from data-objects"""
 
-from ._table import difference, frequencies, melt, melt_ndvar, stats, repmeas
+from ._table import (
+    cast_to_ndvar,
+    difference,
+    frequencies,
+    melt,
+    melt_ndvar,
+    repmeas,
+    stats,
+)